impl admin manage user

+++ src/main/java/com/service/hotel_booking/repositories/UserRepository.java
@@ -6,10 +6,8 @@ import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
-import java.util.Optional;
-
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer>, JpaSpecificationExecutor<User> {
+public interface UserRepository extends JpaRepository<User, Long>, JpaSpecificationExecutor<User> {
 
     @Query("SELECT u FROM User u WHERE u.email = :email")
     User findByEmail(String email);
